requirements-api.txt
# FastAPI Service Dependencies

# Web Framework
fastapi==0.121.3
uvicorn[standard]==0.38.0
pydantic==2.12.4
pydantic-settings==2.12.0

# MLflow Integration
mlflow==2.10.2
boto3==1.34.149  # For MinIO/S3 artifacts

# ML/Data Processing
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.2

# Prometheus Metrics
prometheus-client==0.21.0

# NLP (for text preprocessing)
nltk==3.9.2
regex==2025.11.3

# Utilities
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==1.4.50
